college_predictor_backend/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
pandas==2.0.2
python‐multipart==0.0.6
tabula-py>=2.7.0
